requirements.txt
aioredis==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.15.1
channels==3.0.4
channels-redis==2.4.2
charset-normalizer==3.3.2
constantly==15.1.0
cryptography==41.0.3
daphne==3.0.2
Django==4.0
django-cors-headers==4.2.0
django-eventstream==4.5.1
django-grip==3.5.0
django-restframework==0.0.1
djangorestframework==3.14.0
gripcontrol==4.2.0
hiredis==2.2.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
MarkupSafe==2.1.5
msgpack==0.6.2
pubcontrol==3.5.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
pytz==2023.3
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.2.1
Werkzeug==3.0.3
zope.interface==6.0
